requirements.txt
Django>=1.8
djangorestframework>=2.1.15
msgpack>=1.0.0